Add CompressionOptions to the compressed secondary cache
